add filtering to trace2receiver
Define filtering mechanism in the receiver component
to limit the amount of data that we generate for
some Git commands.

We can drop all telemetry for certain commands or
from certain worktrees.  We can generate summary
data, process-level data, or verbose data for commands.

We define a set of builtin "detail levels" and create
a mechanism to allow custom rulesets to be defined.

Signed-off-by: Jeff Hostetler <jeffhostetler@github.com>