filter_settings_test: unit tests for filtering
Unit tests to verify that we parse the `filter.yml` and
`ruleset.yml` files and that `computeDetailLevel()` correctly
computes the intended detail level using the various defaults,
ruleset_key, nickname, and per-ruleset command maps.

We are only concerned with selecting the correct detail level.
We are not concerned with the detail level's effect on the
generated OTLP.

Signed-off-by: Jeff Hostetler <jeffhostetler@github.com>