reject_client: reject fsmonitor--daemon clients
Create framework for rejecting long-running daemon clients.
Reject all fsmonitor--daemon commands.

Background commands like `git fsmonitor--daemon run` could run
for days/months.  We do not want to hold a connection to them
for the duration and, since we do not send OTLP data until the
client exits, accumulate potentially unlimited telemetry data
from them.

So drop the connection as soon as we detect it is a daemon
process.

NOTE: Since `git fsmonitor--daemon` only sends `cmd_name` (the
verb) and not the `cmd_mode` event, we cannot distinguish
between `run` and `start`/`stop`.  So we just drop them all.

There are other daemon processes that we will also need to
consider, such as `git daemon` and the bundle server.  But
since those are much less likely to be automatically started
and running on a desktop machine, I'll consider them later.

Signed-off-by: Jeff Hostetler <jeffhostetler@github.com>