use the process span id as the parent in child-event spans
Earlier I had set the parent span id of child-event spans
to the span id of the thread-span (associated with the
thread-start and thread-exit events). This helped when
displaying distributed traces because the child span was
closer to the region span of the thread.

However, when the detail level is set to `dl:process` we
do not emit the thread nor region spans, but we do emit
the child span. Therefore, the child span was disconnected.
This was seen in SigNoz as an error in SigNoz. Azure did
not surface this.

I changed the child's parent to be the span id of the
process, so that it will always be connected.

Child spans are a bit of an anomaly anyway because the
the child process doesn't know about any of the nested
thread/region/child spans in the parent process.  The
child only knows about the SID of the parent process and
will use the parent's SID to generate the same TraceID
and span id for the parent process.

So the result here is that the child span will appear
closer to the child's process span, which is probably good
thing.

Signed-off-by: Jeff Hostetler <jeffhostetler@github.com>