rcvr_namedpipe: set explicit ACL on named pipe
If the call to create the server side of the named pipe does not
specify an ACL the pipe code uses `RtlDefaultNpAcl()` to create
a default ACL for the pipe:

    RW NT AUTHORITY\SYSTEM
    RW BUILTIN\Administrators
    R  Everyone
    R  NT AUTHORITY\ANONYMOUS LOGON

If the current user is not a member of `Administrators` then they
won't be able open a client side pipe for writing to send data to
the service.  Also, we should open it up in case there are multiple
user accounts on the system.

Add an ACL to the pipe to allow that.

    RW NT AUTHORITY\SYSTEM
    RW BUILTIN\Administrators
    RW CREATOR OWNER
    RW Everyone

Also, make a slight change to the Microsoft/go-winio API to reduce
confusion. Rename the PipeConfig.SecurityDescriptor field to better
indicate that it is an SDDL rather than a binary structure.  Pass
the SDDL to private routines rather than a `sd []byte` and avoid
a lot of `unsafe.Pointer` casting.

Signed-off-by: Jeff Hostetler <jeffhostetler@github.com>