Docs: add service.telemetry.metrics.level:none to examples
The OpenTelemetry Collector now defaults opening a Prometheus Metrics
port (http://localhost:8888/) when the service starts.  We are not
interested in exposing this data.

Also, it can cause the collector to fail to start because the port is
already in use, such as when another collector instance is already
running. This can happen by accident, for example, when manually
stopping and restarting the service to re-load the config.yml.

Signed-off-by: Jeff Hostetler <jeffhostetler@github.com>