rcvr_unixsocket: detect/handle theft of unix domain socket
Teach the unix domain socket code to detect and handle cases where the
socket in the filesystem is deleted by another process (such as when
another collector daemon instance is started) and gracefully shutdown.

Unix domain sockets have several problems:

(1) If a process fails to delete the socket when it is exiting, the
socket will persist in the filesystem in a "dead" state.  Future
daemons will need to unlink() it before they can create a new socket
in its place.

(2) Future daemons cannot (easily) tell if a socket in the filesystem
is alive or dead.  Typically, services just assume the worst and
delete the socket before creating a new one -- without worrying about
the problems that may cause.

(3) A live socket can be deleted by another process by just calling
unlink() on it. The existing/working daemon does not see any errors,
since Unix just updates the ref-count on the inode and disassociates
the pathname from the socket. So the existing daemon will blindly
continue to listen on a socket that no one can connect to, since the
name was stolen. (The original socket will actually be deleted when
the existing daemon closes the last file descriptor on it in typical
Unix fashion.)

We fix the above problems by:

(A) Having the listener-loop helper thread periodically lstat() the
socket and watch for it to either disappear (such as when another
process deletes the socket) or for the inode to change (such as when
another daemon starts up and deletes-and-recreates the socket).
In either case, we know that the socket we are listening on is no
longer associated with the pathname.

(B) During a shutdown (graceful or otherwise) do not unlink() the
socket if the inode has changed. The normal GOLANG unixsocket code
defaults to closing the file descriptor and unlinking the socket.
If we can detect that another daemon has started up and stolen our
socket pathname, then we don't want to be vindictive and delete their
socket as we shutdown.

Signed-off-by: Jeff Hostetler <jeffhostetler@github.com>