Emit repo nickname as process span attribute
If we have a repo nickname value present in the trace2.param.set, then
emit this as a process span attribute directly "trace2.repo.nickname".
This will make it easier to make decisions on routing of these traces in
the collector, without having to resort to parsing the trace2.param.set
JSON object.

Signed-off-by: Matthew John Cheetham <mjcheetham@outlook.com>