Update Windows CI runner
The `windows-2019` agent image has been removed. We move to
`windows-latest` to avoid this in the future. We should be testing on
the latest Windows as we go forward since we want to make sure our
forked copy of the Windows named pipe IO handling continues to work
correctly with new Windows versions.

Signed-off-by: Matthew John Cheetham <mjcheetham@outlook.com>